Created MySQL Database Schema for Password Manager
The is the schema for the database to hold user information and passwords. It creates two tables, one for user info. and the other for all of the passwords. Each password has its own id and has an id for the user it is connected to. For the users, they also have an id and other info. like their master pass, pass hash, and token.