Made server set cookies instead of client
Made it so that the token cookies are sent using the 'Set-Cookie' header instead of being sent with JSON and having the client-side set them. Also made a logout request that works with HTTP requests. It will remove the token cookies from the client, but removing the refresh token from the database and adding the access token to a denylist still need to be implemented.