Updated files
Fixed typos and added authorization token to privileged.html. Also changed auth_server.js to send the user directly to the privileged page instead of sending a temporary page for redirecting, because this caused issues with the JavaScript. Adding passwords to the database should now work. Getting passwords from the database does work, but they are not being added to the table.